Add send_email_notification_with_file_attached_via_csv  This method is to be used by send file via ui tests to send an email notification with a file attached via csv.  The upload_csv function that is used by other tests did not work for the send file via ui flow and resulted in test failures.
