+++ tests/test_utils.py
@@ -43,13 +43,8 @@ from tests.pages import (
     YourServicesPage,
 )
 from tests.pages.pages import (
-    AddFileToEmailTemplatePage,
-    ChangeLinkTextForEmailFilePage,
     ManageEmailTemplateFilePage,
-    ManageFilesForEmailTemplatePage,
     RenameLetterTemplatePage,
-    ViewEmailTemplatePage, ChangeRentionPeriodForEmailFilePage, EmailConfirmationSettingForEmailFilePage,
-    SendEmailPreviewPage,
 )
 
 logging.basicConfig(filename=f"./logs/test_run_{datetime.now(UTC)}.log", level=logging.INFO)
@@ -650,113 +645,16 @@ def pdf_page_has_text(pdf_page, expected_text, normalise_whitespace=True):
     return expected_text in page_text
 
 
-def create_an_email_template_and_attach_a_file(driver, file_name, template_name, content):
-    go_to_templates_page(driver)
-    template_id = create_email_template(driver, name=template_name, content=content, has_unsubscribe_link=True)
-    add_file_to_email_template(driver, file_name, template_name)
-    return template_id
-
-
-def add_file_to_email_template(driver, file_name, template_name):
-    view_email_template_page = ViewEmailTemplatePage(driver)
-    assert view_email_template_page.get_h1_text() == template_name
-    view_email_template_page.click_attach_files_button()
-    add_file_to_email_template_page = AddFileToEmailTemplatePage(driver)
-    assert add_file_to_email_template_page.get_h1_text() == "Add a file"
-    add_a_file_page = AddFileToEmailTemplatePage(driver)
-    assert add_a_file_page.visible_choose_file_button().is_displayed()
-    file_path = f"tests/test_files/{file_name}"
-    os_file_path = os.path.join(os.getcwd(), file_path)
-    add_a_file_page.upload_file(os_file_path)
-    manage_a_file_page = ManageEmailTemplateFilePage(driver)
-    assert manage_a_file_page.get_h1_text() == file_name
-    manage_a_file_page.click_add_to_template()
-
-
-def assert_file_has_been_attached_to_email_template(driver, template_name):
-    view_email_template_page = ViewEmailTemplatePage(driver)
-    assert view_email_template_page.get_h1_text() == template_name
-    assert view_email_template_page.get_file_added_count_text() == "1 file added"
-
-
 def delete_file_from_email_template_via_manage_files_page(driver):
     manage_a_file_page = ManageEmailTemplateFilePage(driver)
     manage_a_file_page.click_remove_file_link()
     manage_a_file_page.click_remove_file_dialog_button()
 
 
-def go_to_file_management_page_from_email_template_preview(driver, file_name, view_email_template_page):
-    view_email_template_page.click_manage_files_button()
-    manage_files_page = ManageFilesForEmailTemplatePage(driver)
-    assert manage_files_page.get_h1_text() == "Manage files"
-    manage_files_page.click_manage_link(file_name)
-    manage_a_file_page = ManageEmailTemplateFilePage(driver)
-    assert manage_a_file_page.get_h1_text() == file_name
-    return manage_a_file_page, manage_files_page
-
-
-def change_email_template_file_link_text(driver, manage_a_file_page, link_text_label, new_link_text, file_name):
-    default_link_text = manage_a_file_page.get_file_setting_value(link_text_label)
-    assert default_link_text == "Not set"
-    manage_a_file_page.click_change_file_setting(link_text_label)
-    change_link_text_page = ChangeLinkTextForEmailFilePage(driver)
-    assert change_link_text_page.get_h1_text() == "Add link text"
-    change_link_text_page.fill_in_link_text(new_link_text)
-    change_link_text_page.click_continue_button()
-    return change_link_text_page
-
-
-def change_email_template_retention_period(change_link_text_page, driver, manage_a_file_page, new_retention_period,
-                                           retention_period_label):
-    default_retention_period_text = manage_a_file_page.get_file_setting_value(retention_period_label)
-    assert default_retention_period_text == "26 weeks after sending\n        (about 6 months)"
-    manage_a_file_page.click_change_file_setting(retention_period_label)
-    assert change_link_text_page.get_h1_text() == "How long the file is available"
-    change_retention_period = ChangeRentionPeriodForEmailFilePage(driver)
-    change_retention_period.fill_in_retention_period(new_retention_period)
-    change_retention_period.click_continue_button()
-
-
-def change_email_template_email_confirmation(driver, email_confirmation_label, manage_a_file_page,
-                                             new_confirmation_label_choice):
-    confirmation_label_choice = manage_a_file_page.get_file_setting_value(email_confirmation_label)
-    assert confirmation_label_choice == "Yes"
-    manage_a_file_page.click_change_file_setting(email_confirmation_label)
-    email_confirmation_page = EmailConfirmationSettingForEmailFilePage(driver)
-    assert email_confirmation_page.get_h1_text() == "Ask recipient for their email address"
-    email_confirmation_page.select_email_confirmation_option(new_confirmation_label_choice)
-    email_confirmation_page.click_continue_button()
-
-
-def go_back_to_email_template_from_file_management_page(file_name, manage_a_file_page, template_name,
-                                                        view_email_template_page):
-    assert manage_a_file_page.get_h1_text() == file_name
-    manage_a_file_page.click_back_link()
-    assert manage_a_file_page.get_h1_text() == "Manage files"
-    manage_a_file_page.click_back_link()
-    assert view_email_template_page.get_h1_text() == template_name
-
-
-def send_email_notification_with_file_attached_to_one_recipient(driver, view_email_template_page, template_name):
-    view_email_template_page.click_send()
-    set_sender_page = SendSetSenderPage(driver)
-    set_sender_page.wait_until_current()
-    assert set_sender_page.get_h1_text() == "Where should replies come back to?"
-    set_sender_page.click_continue_button()
-    send_to_one_recipient_page = SendOneRecipientPage(driver)
-    assert send_to_one_recipient_page.get_h1_text() == f"Send ‘{template_name}’"
-    send_to_one_recipient_page.send_to_myself("email")
-    preview_send_one_recepient_page = SendEmailPreviewPage(driver)
-    assert preview_send_one_recepient_page.get_h1_text() == f"Preview of ‘{template_name}’"
-    preview_send_one_recepient_page.click_send_button()
-
-
-def go_to_email_template_preview_page_from_a_document_download_pages(driver, template_name,
-                                                                   document_download_page):
+def go_to_email_template_preview_page_from_a_document_download_pages(driver, template_name, document_download_page):
     base_url = config["notify_admin_url"]
     service_template_page_url = f"{base_url}/services/{config['service']['id']}/templates"
     document_download_page.get(service_template_page_url)
     templates_pages = ShowTemplatesPage(driver)
     assert templates_pages.get_h1_text() == "Templates"
     templates_pages.click_template_by_link_text(template_name)
-
